fix(audit): pass 5 — discard regression, plural placeholder, test split
load()'s new revert gate broke the primary Discard path. discardAllAsync
dispatches this page's async revert first, which sets the in-flight flag, so
by the time the settings domain called load() the revert refused, load()
read that as "not clean", and every dirty badge stayed lit after a discard
that had in fact succeeded. A refusal has two causes and they mean opposite
things: the worker owning the restore is the restore proceeding normally,
while a failed one is a page that really is still dirty. load() tells them
apart now.

The saved-sets count used i18np with %1, and this context substitutes only
%n, so the header rendered the literal "%1 sets". The async-discard failure
message had the same problem in the other direction, spelling its plural as
"file(s)" where %n lets each locale pick its own form.

The page-controller toasts were wired twice while a shader browser page was
open, once by the shell and once by the page's own bridge. The shell is the
single owner now, and the snapping bridge it was missing is wired there too,
so nothing is dropped.

Enter did nothing on a focused shader card: AbstractButton handles Space and
only Space, so Return and Enter had to be wired, and they are the keys a user
actually reaches for.

The card body's opacity was animated imperatively on top of its own binding.
A property animation does not drop the binding, so flipping the master toggle
while the card was shut re-evaluated it out from under the animation and the
next expand faded from a value to itself. Height and opacity are both bound to
the collapse progress now, so they move together by construction and only one
property is animated at all.

Enter also submits the save form, through the button rather than around it, so
it cannot step past the overwrite prompt.

test_decoration_sets had grown to 851 lines, past the project's cap and past
what its own header claims it respects. It is split into the round-trip half
and the refusal half, sharing one definition of the sandbox guard. Export's
failure branches and the stem-listed-row case canUseSetName exists for were
the last uncovered paths and are now pinned.
