+++ tests/unit/settings/test_decoration_sets.cpp
@@ -49,6 +49,7 @@
 #include <QJsonDocument>
 #include <QJsonObject>
 #include <QStandardPaths>
+#include <QStringList>
 #include <QTemporaryDir>
 #include <QUrl>
 
@@ -56,65 +57,11 @@
 #include "phosphor_i18n.h"
 #include "settings/decorationpagecontroller.h"
 #include "settings/shadersetstore.h"
+#include "../helpers/DecorationSetHelpers.h"
 #include "../helpers/SetRowHelpers.h"
 #include "../helpers/TreeStubSettings.h"
 
 using namespace PlasmaZones;
-
-namespace {
-
-/// Absolute path to the decoration-sets directory the store writes to,
-/// recomputed the way DecorationPageController does. Valid only under
-/// QStandardPaths test mode (see initTestCase).
-QString decorationSetsDir()
-{
-    const QString base = QStandardPaths::writableLocation(QStandardPaths::GenericDataLocation);
-    return QDir::cleanPath(base + ConfigDefaults::userDecorationSetsSubdir());
-}
-
-/// Write @p root to @p path as a hand-crafted set file.
-void writeSetFile(const QString& path, const QJsonObject& root)
-{
-    QVERIFY(QDir().mkpath(QFileInfo(path).absolutePath()));
-    QFile f(path);
-    QVERIFY(f.open(QIODevice::WriteOnly));
-    const QByteArray bytes = QJsonDocument(root).toJson();
-    QCOMPARE(f.write(bytes), static_cast<qint64>(bytes.size()));
-    f.close();
-}
-
-/// A minimal valid decoration-set payload covering `window.tiled`.
-QJsonObject validSetPayload(const QString& name, int version = 1)
-{
-    QJsonObject profile;
-    profile.insert(QStringLiteral("chain"), QJsonArray{QStringLiteral("glow")});
-    QJsonObject entry;
-    entry.insert(QStringLiteral("path"), QStringLiteral("window.tiled"));
-    entry.insert(QStringLiteral("profile"), profile);
-
-    QJsonObject root;
-    root.insert(QStringLiteral("name"), name);
-    root.insert(QStringLiteral("version"), version);
-    root.insert(QStringLiteral("overrides"), QJsonArray{entry});
-    return root;
-}
-
-/// Recursively clear the sets directory, but ONLY inside the QStandardPaths
-/// sandbox. The guard lives here rather than in initTestCase because QtTest
-/// still runs cleanupTestCase after initTestCase fails: a future edit that
-/// dropped setTestModeEnabled would otherwise delete the user's real saved sets
-/// on its way out.
-void wipeSetsDir()
-{
-    const QString dir = decorationSetsDir();
-    if (!dir.contains(QLatin1String("qttest"))) {
-        qWarning("refusing to wipe a sets directory outside the test sandbox");
-        return;
-    }
-    QDir(dir).removeRecursively();
-}
-
-} // namespace
 
 class TestDecorationSets : public QObject
 {
@@ -146,169 +93,6 @@ private Q_SLOTS:
     }
 
     // ─── Save / list / apply / remove ───────────────────────────────────────
-
-    /// fromJson silently drops unknown and wrong-typed keys, so an entry whose
-    /// chain is a string parses to an all-inherit profile. Staging that would make
-    /// the surface READ as overridden while changing nothing the user can see.
-    void applySet_rejectsAnEntryWhoseProfileEngagesNothing()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        // A well-formed entry on a REAL surface, whose profile parses to nothing:
-        // `chain` is a string where an array belongs.
-        QJsonObject entry;
-        entry.insert(QStringLiteral("path"), QStringLiteral("window.tiled"));
-        entry.insert(QStringLiteral("profile"), QJsonObject{{QStringLiteral("chain"), QStringLiteral("glow")}});
-        QJsonObject root;
-        root.insert(QStringLiteral("name"), QStringLiteral("hollow"));
-        root.insert(QStringLiteral("version"), 1);
-        root.insert(QStringLiteral("overrides"), QJsonArray{entry});
-        writeSetFile(decorationSetsDir() + QStringLiteral("/hollow.json"), root);
-
-        QVERIFY2(!sets->applySet(QStringLiteral("hollow")),
-                 "an override that engages no field must be refused, not staged as a phantom");
-        QVERIFY2(!c.hasOverride(QStringLiteral("window.tiled")), "and nothing may be written to the tree");
-    }
-
-    /// Same rule for the baseline: one that parses to all-inherit would wipe the
-    /// user's global default for no gain.
-    void applySet_rejectsABaselineThatEngagesNothing()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        c.setChain(QString(), QStringList{QStringLiteral("border")});
-
-        QJsonObject entry;
-        entry.insert(QStringLiteral("path"), QStringLiteral("window.tiled"));
-        entry.insert(QStringLiteral("profile"),
-                     QJsonObject{{QStringLiteral("chain"), QJsonArray{QStringLiteral("glow")}}});
-        QJsonObject root;
-        root.insert(QStringLiteral("name"), QStringLiteral("hollowbase"));
-        root.insert(QStringLiteral("version"), 1);
-        root.insert(QStringLiteral("overrides"), QJsonArray{entry});
-        // Non-empty object, but every key is wrong-typed, so it parses to nothing.
-        root.insert(QStringLiteral("baseline"), QJsonObject{{QStringLiteral("chain"), 7}});
-        writeSetFile(decorationSetsDir() + QStringLiteral("/hollowbase.json"), root);
-
-        QVERIFY2(!sets->applySet(QStringLiteral("hollowbase")), "a baseline that engages no field must be refused");
-        QCOMPARE(c.chainAt(QString()), (QStringList{QStringLiteral("border")}));
-    }
-
-    /// A present-but-non-array `overrides` used to read as "no overrides", so a set
-    /// could import and apply with everything it claimed silently dropped.
-    void applySet_rejectsNonArrayOverrides()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        QJsonObject root;
-        root.insert(QStringLiteral("name"), QStringLiteral("bent"));
-        root.insert(QStringLiteral("version"), 1);
-        root.insert(QStringLiteral("overrides"), QStringLiteral("not an array"));
-        root.insert(QStringLiteral("baseline"),
-                    QJsonObject{{QStringLiteral("chain"), QJsonArray{QStringLiteral("border")}}});
-        writeSetFile(decorationSetsDir() + QStringLiteral("/bent.json"), root);
-
-        QVERIFY2(!sets->applySet(QStringLiteral("bent")),
-                 "a malformed overrides field must refuse the whole set, not apply the baseline alone");
-    }
-
-    /// Every mutator resolves a row by name to slugify(name) + ".json", so a file
-    /// whose stem does not match cannot be applied, renamed or deleted. Listing it
-    /// would show the user a row they can never act on.
-    void availableSets_skipsAnUnaddressableFile()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        // Hand-placed: the stem is "My Set", but "My Set" slugifies to "my-set".
-        QJsonObject root;
-        root.insert(QStringLiteral("name"), QStringLiteral("My Set"));
-        root.insert(QStringLiteral("version"), 1);
-        QJsonObject entry;
-        entry.insert(QStringLiteral("path"), QStringLiteral("window.tiled"));
-        entry.insert(QStringLiteral("profile"),
-                     QJsonObject{{QStringLiteral("chain"), QJsonArray{QStringLiteral("glow")}}});
-        root.insert(QStringLiteral("overrides"), QJsonArray{entry});
-        writeSetFile(decorationSetsDir() + QStringLiteral("/My Set.json"), root);
-
-        QTest::ignoreMessage(QtWarningMsg, QRegularExpression(QStringLiteral("skipping")));
-        QVERIFY2(rowFor(sets, QStringLiteral("My Set")).isEmpty(), "an unaddressable file must not be listed");
-        QVERIFY2(rowFor(sets, QStringLiteral("my-set")).isEmpty(), "and not under its slug either");
-    }
-
-    /// canUseSetName is the Ok-gate for the rename dialog, and the ONLY thing
-    /// standing between the user and a dismissed dialog that silently drops their
-    /// description edit. It has to agree with updateSet exactly.
-    void canUseSetName_matchesWhatUpdateSetAccepts()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        c.setChain(QStringLiteral("window.tiled"), QStringList{QStringLiteral("glow")});
-        QVERIFY(sets->saveCurrentAsSet(QStringLiteral("Mine"), QString()));
-        QVERIFY(sets->saveCurrentAsSet(QStringLiteral("Other"), QString()));
-
-        QVERIFY2(!sets->canUseSetName(QString(), QStringLiteral("Mine")), "an empty name is refused");
-        QVERIFY2(!sets->canUseSetName(QStringLiteral("   "), QStringLiteral("Mine")), "whitespace only is refused");
-        QVERIFY2(!sets->canUseSetName(QStringLiteral("!!!"), QStringLiteral("Mine")),
-                 "a name that slugifies to nothing is refused");
-        QVERIFY2(!sets->canUseSetName(QStringLiteral("Other"), QStringLiteral("Mine")),
-                 "a collision with a DIFFERENT set is refused");
-        QVERIFY2(sets->canUseSetName(QStringLiteral("Mine"), QStringLiteral("Mine")),
-                 "keeping your own name is a description-only edit, not a self-collision");
-        QVERIFY2(sets->canUseSetName(QStringLiteral("Fresh"), QStringLiteral("Mine")), "a free name is accepted");
-
-        // Now prove the predicate agrees with the mutator on every one of those.
-        QVERIFY(!sets->updateSet(QStringLiteral("Mine"), QString(), QString()));
-        QVERIFY(!sets->updateSet(QStringLiteral("Mine"), QStringLiteral("!!!"), QString()));
-        QVERIFY(!sets->updateSet(QStringLiteral("Mine"), QStringLiteral("Other"), QString()));
-        QVERIFY(sets->updateSet(QStringLiteral("Mine"), QStringLiteral("Mine"), QStringLiteral("desc only")));
-        QVERIFY(sets->updateSet(QStringLiteral("Mine"), QStringLiteral("Fresh"), QString()));
-    }
-
-    /// A write that cannot land must fail loudly and list nothing. Provoked with a
-    /// directory in the destination's place rather than by revoking permissions, so
-    /// it runs as root too. CLAUDE.md's documented build flow is Docker, which runs
-    /// as root, and a permissions-based test would silently QSKIP there.
-    ///
-    /// This does NOT pin the snapshot rollback: decoration wires no fileSnapshot
-    /// hook, so nothing is ever staged here. The rollback is pinned on the motion
-    /// side, which is the domain that stages.
-    void failedSetWriteToastsAndListsNothing()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        // A DIRECTORY where the set file wants to be: QSaveFile cannot commit
-        // over it, so the write fails after the snapshot has been staged.
-        const QString destination = decorationSetsDir() + QStringLiteral("/doomed.json");
-        QVERIFY(QDir().mkpath(destination));
-        QVERIFY(QFileInfo(destination).isDir());
-
-        c.setChain(QStringLiteral("window.tiled"), QStringList{QStringLiteral("glow")});
-        QSignalSpy toastSpy(sets, &ShaderSetStore::toastRequested);
-        QTest::ignoreMessage(QtWarningMsg, QRegularExpression(QStringLiteral("could not write")));
-        // overwrite=true: the destination "exists" (it is the directory), so the
-        // collision check would otherwise refuse before the write is even tried,
-        // and the write is what this test is about.
-        QVERIFY2(!sets->saveCurrentAsSet(QStringLiteral("Doomed"), QString(), /*overwrite=*/true),
-                 "a save whose write cannot land must fail");
-        QCOMPARE(toastSpy.count(), 1);
-        QCOMPARE(toastSpy.first().first().toString(), PhosphorI18n::tr("Could not write the set to disk."));
-
-        // The directory is untouched, and no half-written set is listed.
-        QVERIFY(QFileInfo(destination).isDir());
-        QVERIFY2(rowFor(sets, QStringLiteral("Doomed")).isEmpty(), "no half-written set may be listed");
-    }
 
     /// Apply MERGES: a surface the set does not cover keeps whatever chain it
     /// has. The file header pins this and the motion side has the same test, but
@@ -533,269 +317,9 @@ private Q_SLOTS:
         QCOMPARE(c.chainAt(QStringLiteral("window.tiled")), (QStringList{QStringLiteral("glow")}));
     }
 
-    /// Import validates against the surface taxonomy, so a file carrying an
-    /// unknown path (a motion set, say) is refused at the boundary rather than
-    /// landing on disk and failing later at apply time.
-    void decorationSets_importRejectsForeignPayload()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        QTemporaryDir dir;
-        QVERIFY(dir.isValid());
-        const QString foreign = dir.filePath(QStringLiteral("foreign.json"));
-
-        QJsonObject entry;
-        entry.insert(QStringLiteral("path"), QStringLiteral("window.appearance.open")); // a motion path
-        // A REAL profile: an empty one would be refused by the empty-profile
-        // rule, and the path-taxonomy guard under test would never run.
-        entry.insert(QStringLiteral("profile"),
-                     QJsonObject{{QStringLiteral("chain"), QJsonArray{QStringLiteral("glow")}}});
-        QJsonObject root;
-        root.insert(QStringLiteral("name"), QStringLiteral("Foreign"));
-        root.insert(QStringLiteral("version"), 1);
-        root.insert(QStringLiteral("overrides"), QJsonArray{entry});
-        writeSetFile(foreign, root);
-
-        QVERIFY2(!sets->importSet(foreign), "a set whose paths are not decoration surfaces must be refused");
-        QVERIFY(sets->availableSets().isEmpty());
-    }
-
     // ─── Version gate ───────────────────────────────────────────────────────
 
-    /// A set written by a NEWER build may carry fields this build drops on
-    /// parse, so applying it would commit a silently truncated look. Both apply
-    /// and import must refuse it. A non-numeric version is malformed and is
-    /// refused for the same reason.
-    void decorationSets_refusesNewerFormatVersion()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        // A future-format set already sitting in the sets dir.
-        writeSetFile(decorationSetsDir() + QStringLiteral("/from-the-future.json"),
-                     validSetPayload(QStringLiteral("from-the-future"), 2));
-        QVERIFY2(!sets->applySet(QStringLiteral("from-the-future")), "a set from a newer build must not be applied");
-        QVERIFY2(!c.hasOverride(QStringLiteral("window.tiled")), "the refused set must not have been staged");
-
-        // The same set arriving through import.
-        QTemporaryDir dir;
-        QVERIFY(dir.isValid());
-        const QString newer = dir.filePath(QStringLiteral("newer.json"));
-        writeSetFile(newer, validSetPayload(QStringLiteral("Newer"), 2));
-        QVERIFY2(!sets->importSet(newer), "a set from a newer build must not be imported");
-
-        // A present-but-non-numeric version is malformed, not "version 1".
-        const QString malformed = dir.filePath(QStringLiteral("malformed.json"));
-        QJsonObject root = validSetPayload(QStringLiteral("Malformed"));
-        root.insert(QStringLiteral("version"), QStringLiteral("1"));
-        writeSetFile(malformed, root);
-        QVERIFY2(!sets->importSet(malformed), "a non-numeric version must be refused");
-
-        // A non-integral version is malformed too. QJsonValue::toInt() would
-        // hand back the default for it, silently reading 1.5 as version 1.
-        const QString fractional = dir.filePath(QStringLiteral("fractional.json"));
-        QJsonObject frac = validSetPayload(QStringLiteral("Fractional"));
-        frac.insert(QStringLiteral("version"), 1.5);
-        writeSetFile(fractional, frac);
-        QVERIFY2(!sets->importSet(fractional), "a non-integral version must be refused");
-
-        // A set with no version at all reads as the current format.
-        const QString versionless = dir.filePath(QStringLiteral("versionless.json"));
-        QJsonObject noVersion = validSetPayload(QStringLiteral("Versionless"));
-        noVersion.remove(QStringLiteral("version"));
-        writeSetFile(versionless, noVersion);
-        QVERIFY2(sets->importSet(versionless), "a versionless set must read as the current format");
-    }
-
     // ─── Refusals ───────────────────────────────────────────────────────────
-
-    /// Saving onto an existing name destroys that set, and decoration wires no
-    /// file-snapshot hook, so nothing could restore it. An UNCONFIRMED save is
-    /// therefore refused (with a toast) and leaves the stored set intact.
-    /// Re-saving with the user's consent (overwrite=true) is how they update a
-    /// set after tweaking their look, so that path must still work.
-    void saveDecorationSet_refusesUnconfirmedOverwriteAndHonoursConsent()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        c.setChain(QStringLiteral("window.tiled"), QStringList{QStringLiteral("glow")});
-        QVERIFY(sets->saveCurrentAsSet(QStringLiteral("Taken"), QStringLiteral("first")));
-        QCOMPARE(sets->existingSetName(QStringLiteral("Taken")), QStringLiteral("Taken"));
-        QVERIFY(sets->existingSetName(QStringLiteral("Not Taken")).isEmpty());
-
-        // A different look, saved under the same name, WITHOUT consent.
-        c.setChain(QStringLiteral("window.tiled"), QStringList{QStringLiteral("border")});
-        QSignalSpy toastSpy(sets, &ShaderSetStore::toastRequested);
-        QVERIFY2(!sets->saveCurrentAsSet(QStringLiteral("Taken"), QStringLiteral("second")),
-                 "an unconfirmed overwrite must be refused");
-        QCOMPARE(toastSpy.count(), 1);
-
-        // The stored set is intact — same description, same payload.
-        const QVariantMap row = rowFor(sets, QStringLiteral("Taken"));
-        QVERIFY(!row.isEmpty());
-        QCOMPARE(row.value(QStringLiteral("description")).toString(), QStringLiteral("first"));
-        QVERIFY(sets->applySet(QStringLiteral("Taken")));
-        QCOMPARE(c.chainAt(QStringLiteral("window.tiled")), (QStringList{QStringLiteral("glow")}));
-
-        // Now WITH consent: the set is re-pointed at the new look.
-        c.setChain(QStringLiteral("window.tiled"), QStringList{QStringLiteral("border")});
-        QVERIFY2(sets->saveCurrentAsSet(QStringLiteral("Taken"), QStringLiteral("second"), true),
-                 "a confirmed overwrite must be honoured");
-        QCOMPARE(sets->availableSets().size(), 1);
-        const QVariantMap updated = rowFor(sets, QStringLiteral("Taken"));
-        QVERIFY(!updated.isEmpty());
-        QCOMPARE(updated.value(QStringLiteral("description")).toString(), QStringLiteral("second"));
-
-        // The re-saved payload is the NEW look.
-        c.setChain(QStringLiteral("window.tiled"), QStringList{QStringLiteral("glow")});
-        QVERIFY(sets->applySet(QStringLiteral("Taken")));
-        QCOMPARE(c.chainAt(QStringLiteral("window.tiled")), (QStringList{QStringLiteral("border")}));
-    }
-
-    /// applySet is fire-and-forget from QML, so every failure branch must carry
-    /// its reason to the toast or the user just watches nothing happen. Covers
-    /// the three branches reachable here (read / version / validation); the
-    /// apply-failed branch needs a failing ISettings, and the mutation-guard
-    /// branch is motion-only (decoration wires no guard).
-    void applySet_toastsReadVersionAndValidationFailures()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        // Missing file.
-        QSignalSpy missingSpy(sets, &ShaderSetStore::toastRequested);
-        QVERIFY(!sets->applySet(QStringLiteral("nonexistent")));
-        QCOMPARE(missingSpy.count(), 1);
-        QCOMPARE(missingSpy.first().first().toString(),
-                 PhosphorI18n::tr("Could not read the set \"%1\".").arg(QStringLiteral("nonexistent")));
-
-        // Newer format version.
-        writeSetFile(decorationSetsDir() + QStringLiteral("/newer.json"), validSetPayload(QStringLiteral("newer"), 2));
-        QSignalSpy versionSpy(sets, &ShaderSetStore::toastRequested);
-        QVERIFY(!sets->applySet(QStringLiteral("newer")));
-        QCOMPARE(versionSpy.count(), 1);
-        QCOMPARE(
-            versionSpy.first().first().toString(),
-            PhosphorI18n::tr("\"%1\" was written by a newer version of PlasmaZones.").arg(QStringLiteral("newer")));
-
-        // Foreign payload (an event path, not a decoration surface).
-        QJsonObject entry;
-        entry.insert(QStringLiteral("path"), QStringLiteral("window.appearance.open"));
-        // A REAL profile: an empty one would be refused by the empty-profile
-        // rule, and the path-taxonomy guard under test would never run.
-        entry.insert(QStringLiteral("profile"),
-                     QJsonObject{{QStringLiteral("chain"), QJsonArray{QStringLiteral("glow")}}});
-        QJsonObject foreign;
-        foreign.insert(QStringLiteral("name"), QStringLiteral("foreign"));
-        foreign.insert(QStringLiteral("version"), 1);
-        foreign.insert(QStringLiteral("overrides"), QJsonArray{entry});
-        writeSetFile(decorationSetsDir() + QStringLiteral("/foreign.json"), foreign);
-        QSignalSpy validateSpy(sets, &ShaderSetStore::toastRequested);
-        QVERIFY(!sets->applySet(QStringLiteral("foreign")));
-        QCOMPARE(validateSpy.count(), 1);
-        QCOMPARE(validateSpy.first().first().toString(),
-                 PhosphorI18n::tr("\"%1\" does not match this page.").arg(QStringLiteral("foreign")));
-    }
-
-    /// An import carrying no entries would land as a row that applySet then
-    /// refuses. It is named for what it is, rather than falling through to the
-    /// taxonomy message, which would misdescribe an empty file as a foreign one.
-    void decorationSets_importRejectsEmptySet()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        QTemporaryDir dir;
-        QVERIFY(dir.isValid());
-        const QString empty = dir.filePath(QStringLiteral("empty.json"));
-        QJsonObject root;
-        root.insert(QStringLiteral("name"), QStringLiteral("Empty"));
-        root.insert(QStringLiteral("version"), 1);
-        root.insert(QStringLiteral("overrides"), QJsonArray{});
-        writeSetFile(empty, root);
-
-        QSignalSpy toastSpy(sets, &ShaderSetStore::toastRequested);
-        QVERIFY2(!sets->importSet(empty), "a set carrying nothing must be refused");
-        QCOMPARE(toastSpy.count(), 1);
-        // The taxonomy validator would refuse an empty set anyway, so the ONLY
-        // thing the store's dedicated empty check buys is an honest message.
-        // Assert it, or this test passes with that check deleted.
-        QCOMPARE(toastSpy.first().first().toString(), PhosphorI18n::tr("That set is empty."));
-        QVERIFY(sets->availableSets().isEmpty());
-    }
-
-    /// A set file larger than the read cap is refused before it is slurped into
-    /// memory. importSet takes a user-chosen path, so this is a boundary check.
-    void decorationSets_importRejectsOversizedFile()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        QTemporaryDir dir;
-        QVERIFY(dir.isValid());
-        const QString huge = dir.filePath(QStringLiteral("huge.json"));
-
-        // A VALID set, padded past the cap with a giant description. Garbage
-        // bytes would prove nothing: the JSON parser would reject them even
-        // with the cap removed, so the test could not tell the two apart.
-        QJsonObject root = validSetPayload(QStringLiteral("Huge"));
-        root.insert(QStringLiteral("description"), QString(5 * 1024 * 1024, QLatin1Char('x')));
-        writeSetFile(huge, root);
-        QVERIFY(QFileInfo(huge).size() > 4 * 1024 * 1024);
-
-        QVERIFY2(!sets->importSet(huge), "a set file over the size cap must be refused");
-        QVERIFY(sets->availableSets().isEmpty());
-    }
-
-    /// Saving with an empty tree (no baseline, no overrides) is refused: the
-    /// resulting set would be a no-op that applySet then rejects, so it must
-    /// never reach disk.
-    void saveDecorationSet_emptyTreeRejected()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        QSignalSpy setsSpy(sets, &ShaderSetStore::setsChanged);
-        QSignalSpy toastSpy(sets, &ShaderSetStore::toastRequested);
-        QVERIFY2(!sets->saveCurrentAsSet(QStringLiteral("Nothing"), QString()),
-                 "saving an empty decoration tree must be refused");
-        QCOMPARE(setsSpy.count(), 0);
-        QCOMPARE(toastSpy.count(), 1); // the refusal must say why
-        QVERIFY(sets->availableSets().isEmpty());
-    }
-
-    /// saveCurrentAsSet rejects an empty name (would slugify to an empty
-    /// filename).
-    void saveDecorationSet_emptyOrUnusableNameRejected()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-        c.setChain(QString(), QStringList{QStringLiteral("border")});
-
-        QVERIFY(!sets->saveCurrentAsSet(QString(), QString()));
-
-        // A name with nothing a filename can be built from ("!!!" slugifies to
-        // nothing). The Save button only checks for non-blank text, so this
-        // reaches the store and must be refused WITH a reason — otherwise the
-        // user clicks Save and watches nothing happen.
-        QSignalSpy toastSpy(sets, &ShaderSetStore::toastRequested);
-        QVERIFY2(!sets->saveCurrentAsSet(QStringLiteral("!!!"), QString()),
-                 "a name that slugifies to nothing must be refused");
-        QCOMPARE(toastSpy.count(), 1);
-        QCOMPARE(toastSpy.first().first().toString(),
-                 PhosphorI18n::tr("That name cannot be used. Try one with letters or numbers in it."));
-        QVERIFY(sets->availableSets().isEmpty());
-    }
 
     /// applySet validates every entry up-front and rejects the whole set on any
     /// unknown surface path, leaving the current tree untouched (atomic apply —
@@ -829,20 +353,6 @@ private Q_SLOTS:
         QVERIFY2(!c.hasOverride(QStringLiteral("window.snapped")), "applySet wrote partial state from a malformed set");
         // The pre-existing override is untouched.
         QCOMPARE(c.chainAt(QStringLiteral("window.tiled")), (QStringList{QStringLiteral("border")}));
-    }
-
-    /// applySet / removeSet on a name with no file report failure rather than
-    /// crashing or emitting a spurious change.
-    void decorationSets_unknownNameReturnsFalse()
-    {
-        TreeStubSettings settings;
-        DecorationPageController c(nullptr, &settings);
-        ShaderSetStore* sets = c.setsBridge();
-
-        QSignalSpy setsSpy(sets, &ShaderSetStore::setsChanged);
-        QVERIFY(!sets->applySet(QStringLiteral("nonexistent")));
-        QVERIFY(!sets->removeSet(QStringLiteral("nonexistent")));
-        QCOMPARE(setsSpy.count(), 0);
     }
 };
